fix/grpc.footPrintControllerを実装

+++ grpc/repository/foot_print.go
@@ -10,7 +10,7 @@ type FootPrintRepository interface {
 	GetFootPrintsByUid(currentUserUid string) ([]*domain.FootPrint, error)
 	CreateFootPrint(footPrint *domain.FootPrint) error
 	UpdateToAlreadyRead(currentUserUid string) error
-	GetUnreadCount(currentUserUid string) (unreadCount int, err error)
+	GetUnreadCount(currentUserUid string) (unreadCount int64, err error)
 }
 
 type footPrintRepository struct {
@@ -51,7 +51,7 @@ func (r *footPrintRepository) UpdateToAlreadyRead(currentUserUid string) error {
 	return nil
 }
 
-func (r *footPrintRepository) GetUnreadCount(currentUserUid string) (unreadCount int, err error) {
+func (r *footPrintRepository) GetUnreadCount(currentUserUid string) (unreadCount int64, err error) {
 	query := `SELECT count(*) AS unreadCount FROM foot_prints WHERE user_uid = ? AND unread = true`
 	// HACK: Rawだけだと、SQLの結果1行が返されてるのでRowで絞り込み？
 	row := r.db.Raw(query, currentUserUid).Row()
